M8: fix four PostgreSQL defects the live run found
The library's tests run on SQLite, and SQLite accepts what PostgreSQL refuses.
Running the milestone's own journey against PostgreSQL 17.11 found four defects in
its new code, every one of them invisible on SQLite:

* a raw `sqlx::query_scalar` on the PostgreSQL branch handed PostgreSQL a literal
  `?` — `db.sql` rewrites those, but only for the strings that pass through it;
* `?::bigint::boolean` in three places: the bind is an i64, which sqlx sends as
  BIGINT, and PostgreSQL has no bigint-to-boolean cast, so it is `?::int::boolean`;
* the per-page facts query compared `account_id` without `::uuid` and projected
  uuid columns without `::text`, so every listing answered 500;
* `SUM()` of a bigint is `numeric` and will not decode into an i64 — two sites in
  the storage figures, and two more of the same class found elsewhere in the tree:
  the work list's `word_count`, and `public_rating_summary`'s `stars`, which was
  hidden behind its own HAVING clause returning no rows until a work has ratings.

Also wires the reading-status read, which the db layer supported and no route
offered, and adds `scripts/postgres-journey.sh` with a `just journey-postgres`
recipe so the claim is repeatable rather than made once.

47 steps, 0 failures, 0 driver complaints, 0 five-hundreds on PostgreSQL 17.11.

+++ crates/db/src/content.rs
@@ -555,7 +555,7 @@ pub async fn works_for_pseud(db: &Database, pseud: PseudId) -> Result<Vec<OwnedW
                 w.published_at, w.version,
                 (SELECT COUNT(*) FROM chapters c
                   WHERE c.work_id = w.id AND c.deleted_at IS NULL) AS chapter_count,
-                (SELECT COALESCE(SUM(r.word_count), 0)
+                (SELECT COALESCE(SUM(r.word_count), 0)::bigint
                    FROM chapters c
                    JOIN chapter_revisions r ON r.id = c.current_revision_id
                   WHERE c.work_id = w.id AND c.deleted_at IS NULL) AS word_count,
